+++ drift/test/streams_test.dart
@@ -4,7 +4,6 @@ import 'package:drift/drift.dart';
 import 'package:drift/src/runtime/api/runtime_api.dart';
 import 'package:drift/src/runtime/executor/stream_queries.dart';
 import 'package:mockito/mockito.dart';
-import 'package:rxdart/rxdart.dart';
 import 'package:test/test.dart';
 
 import 'data/tables/custom_tables.dart';
@@ -423,32 +422,5 @@ void main() {
   test('stream queries are broadcasts', () {
     expect(db.customSelect('SELECT 1').watch().isBroadcast, isTrue);
     expect(db.customSelect('SELECT 1').watchSingle().isBroadcast, isTrue);
-  });
-
-  test('drift streams can be used with switchMap in rxdart', () async {
-    // Regression test for https://github.com/simolus3/moor/issues/500
-    when(executor.runSelect(any, any)).thenAnswer((i) async {
-      final sql = i.positionalArguments.first as String;
-
-      return [
-        if (sql.contains("'a'")) {'a': 'a'} else {'b': 'b'}
-      ];
-    });
-
-    final a = db
-        .customSelect("select 'a' as a")
-        .map(($) => $.readString('a'))
-        .watchSingle();
-    final b = db
-        .customSelect("select 'b' as b")
-        .map(($) => $.readString('b'))
-        .watchSingle();
-    final c = a.switchMap((_) => b);
-    expect(await a.first, 'a');
-    expect(await a.first, 'a');
-    expect(await b.first, 'b');
-    expect(await b.first, 'b');
-    expect(await c.first, 'b');
-    expect(await c.first, 'b');
   });
 }
